[MTB-007:docs] fix: namespacing isolation. Trello: https://trello.com/c/bBZ5awnv
