+++ lib/ace_config.rb
@@ -5,4 +5,4 @@ require "ace_config/errors"
 require "ace_config/type_map"
 require "ace_config/setting"
 require "ace_config/type_checker"
-require "ace_config/ace_config"
+require "ace_config/configuration"
